Feat: 헤더 카테고리 패널 추가
Resolves: #2
